Implementando métodos

+++ University/Program Language III/Projeto - Sigma Bank/src/produtosBancarios/emprestimos/EmprestimoConsignado.java
@@ -3,7 +3,7 @@ package produtosBancarios.emprestimos;
 public class EmprestimoConsignado extends Emprestimo{
 
     // Diferença na taxa de Juros
-    public EmprestimoConsignado( int qtdParcelas, int valor ) {
+    public EmprestimoConsignado( int qtdParcelas, double valor ) {
         super( qtdParcelas, valor + ( valor * 0.10d ) );
     }
