+++ University/Program Language III/Projeto - Sigma Bank/src/produtosBancarios/emprestimos/EmprestimoPessoal.java
@@ -3,7 +3,7 @@ package produtosBancarios.emprestimos;
 public class EmprestimoPessoal extends Emprestimo {
 
     // Diferença na taxa de Juros
-    public EmprestimoPessoal( int qtdParcelas, int valor ) {
+    public EmprestimoPessoal( int qtdParcelas, double valor ) {
         super( qtdParcelas, valor + (valor * 0.20d) );
     }
 }
